moves the ExceptionHandling module to engine.rb.
- it’s very coupled with rails, and so it felt like maybe it wasn’t the right spot for it anymore.
