+++ lib/teaspoon/exceptions.rb
@@ -18,31 +18,4 @@ module Teaspoon
   class DependencyFailure < Teaspoon::Error; end
   class ThresholdNotMet < Teaspoon::Error; end
   class FileNotWritable < Teaspoon::Error; end
-
-  module ExceptionHandling
-    def self.add_rails_handling
-      return unless Teaspoon.configuration.driver == "phantomjs"
-
-      # debugging should be off to display errors in the suite_controller
-      # Rails.application.config.assets.debug = false
-
-      # we want rails to display exceptions
-      Rails.application.config.action_dispatch.show_exceptions = true
-
-      # override the render exception method in ActionDispatch to raise a javascript exception
-      render_exceptions_with_javascript
-    end
-
-    private
-
-    def self.render_exceptions_with_javascript
-      ActionDispatch::DebugExceptions.class_eval do
-        def render_exception(_env, exception)
-          message = "#{exception.class.name}: #{exception.message}"
-          body = "<script>throw Error(#{[message, exception.backtrace].join("\n").inspect})</script>"
-          [200, { "Content-Type" => "text/html;", "Content-Length" => body.bytesize.to_s }, [body]]
-        end
-      end
-    end
-  end
 end
